Unbundle types
