+++ src/types/Fields.ts
@@ -1,0 +1,18 @@
+export default interface Fields {
+    '${auth_access_token}': string;
+    '${auth_session}': string;
+    '${auth_player_name}': string;
+    '${auth_uuid}': string;
+    '${auth_xuid}': string;
+    '${user_properties}': '{}';
+    '${user_type}': string;
+    '${version_name}'?: string;
+    '${assets_index_name}': string;
+    '${game_directory}': string;
+    '${assets_root}': string;
+    '${game_assets}': string;
+    '${version_type}': string;
+    '${clientid}': string;
+    '${resolution_width}'?: number;
+    '${resolution_height}'?: number;
+}
